test: put existing retarget tests to a separate file
