+++ tests/page/elementhandle-wait-for-element-state.spec.ts
@@ -83,17 +83,6 @@ it('should wait for hidden when detached', async ({ page }) => {
   await promise;
 });
 
-it('should wait for enabled button', async ({ page, server }) => {
-  await page.setContent('<button disabled><span>Target</span></button>');
-  const span = await page.$('text=Target');
-  let done = false;
-  const promise = span.waitForElementState('enabled').then(() => done = true);
-  await giveItAChanceToResolve(page);
-  expect(done).toBe(false);
-  await span.evaluate(span => (span.parentElement as HTMLButtonElement).disabled = false);
-  await promise;
-});
-
 it('should throw waiting for enabled when detached', async ({ page }) => {
   await page.setContent(`<button disabled>Target</button>`);
   const button = await page.$('button');
@@ -101,17 +90,6 @@ it('should throw waiting for enabled when detached', async ({ page }) => {
   await button.evaluate(button => button.remove());
   const error = await promise;
   expect(error.message).toContain('Element is not attached to the DOM');
-});
-
-it('should wait for button with a disabled fieldset', async ({ page }) => {
-  await page.setContent('<fieldset disabled=true><button><span>Target</span></button></div>');
-  const span = await page.$('text=Target');
-  let done = false;
-  const promise = span.waitForElementState('enabled').then(() => done = true);
-  await giveItAChanceToResolve(page);
-  expect(done).toBe(false);
-  await span.evaluate(span => (span.parentElement.parentElement as HTMLFieldSetElement).disabled = false);
-  await promise;
 });
 
 it('should wait for aria enabled button', async ({ page }) => {
@@ -133,17 +111,6 @@ it('should wait for button with an aria-disabled parent', async ({ page }) => {
   await giveItAChanceToResolve(page);
   expect(done).toBe(false);
   await span.evaluate(span => span.parentElement.parentElement.setAttribute('aria-disabled', 'false'));
-  await promise;
-});
-
-it('should wait for disabled button', async ({ page }) => {
-  await page.setContent('<button><span>Target</span></button>');
-  const span = await page.$('text=Target');
-  let done = false;
-  const promise = span.waitForElementState('disabled').then(() => done = true);
-  await giveItAChanceToResolve(page);
-  expect(done).toBe(false);
-  await span.evaluate(span => (span.parentElement as HTMLButtonElement).disabled = true);
   await promise;
 });
