+++ tests/page/page-check.spec.ts
@@ -41,41 +41,6 @@ it('should not uncheck the unchecked box', async ({ page }) => {
   expect(await page.evaluate(() => window['checkbox'].checked)).toBe(false);
 });
 
-it('should check the box by label', async ({ page }) => {
-  await page.setContent(`<label for='checkbox'><input id='checkbox' type='checkbox'></input></label>`);
-  await page.check('label');
-  expect(await page.evaluate(() => window['checkbox'].checked)).toBe(true);
-});
-
-it('should check the box outside label', async ({ page }) => {
-  await page.setContent(`<label for='checkbox'>Text</label><div><input id='checkbox' type='checkbox'></input></div>`);
-  await page.check('label');
-  expect(await page.evaluate(() => window['checkbox'].checked)).toBe(true);
-});
-
-it('should check the box inside label w/o id', async ({ page }) => {
-  await page.setContent(`<label>Text<span><input id='checkbox' type='checkbox'></input></span></label>`);
-  await page.check('label');
-  expect(await page.evaluate(() => window['checkbox'].checked)).toBe(true);
-});
-
-it('should check the box outside shadow dom label', async ({ page }) => {
-  await page.setContent('<div></div>');
-  await page.$eval('div', div => {
-    const root = div.attachShadow({ mode: 'open' });
-    const label = document.createElement('label');
-    label.setAttribute('for', 'target');
-    label.textContent = 'Click me';
-    root.appendChild(label);
-    const input = document.createElement('input');
-    input.setAttribute('type', 'checkbox');
-    input.setAttribute('id', 'target');
-    root.appendChild(input);
-  });
-  await page.check('label');
-  expect(await page.$eval('input', input => input.checked)).toBe(true);
-});
-
 it('should check radio', async ({ page }) => {
   await page.setContent(`
     <input type='radio'>one</input>
